requirement.txt
azure-core
azure-ai-translation-document
azure-storage-blob
